Fix opening mpv for URLs with % chars
For context, when first arg string to `message` contains % it is
treated by `message` as a template string and it expects more args to
come.
Basically, it acts as a `format`. Why? Go ask.
Docs:
https://www.gnu.org/software/emacs/manual/html_node/eintr/message.html

As a result, URLs with % in it were failing to open mpv on click due to error
from `message`:
error: "Not enough arguments for format string"
E.g., for such a link:
[[video:https://us02web.zoom.us/rec/play/ZIwV9OI_nOvLsM4E6qcWC95qzItbgYkySNmpunguzYC7X8amtyiQD1hTCMwZ5ETLAlVRh0Sgo1qmErQa.JamVkhsxNlyjU91t?canPlayFromShare=true&from=share_recording_detail&continueMode=true&componentName=rec-play&originRequestUrl=https%3A%2F%2Fus02web.zoom.us%2Frec%2Fshare%2Fkzc-Rsm_XOBvBMoSttIupf3MO8WmP-tEopTPh7WD9qykRWT1tvBGTrDMsMN8TCLq.qf1jJEwjs8A6xhwX#0:29:47][0:29:47]] yada yada